delete connectionpooler_types.go
